pyproject.toml
[tool.poetry]
name = "argus"
version = "0.1.0"
description = "Argus SDK: Autonomous Business Intelligence Agent"
authors = ["Antigravity <antigravity@google.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langgraph = "^0.2.0"
langchain = "^0.3.0"
langchain-openai = "^0.2.0"
pydantic = "^2.0.0"
sqlglot = "^20.0.0"
sqlalchemy = "^2.0.0"
duckdb = "^1.0.0"
pandas = "^2.0.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.0"
langchain-google-genai = "*"
langchain-anthropic = "*"
pymongo = "^4.0.0"
chromadb = "^0.4.0"
openpyxl = "^3.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
